+++ lib/pavlov/alpha_compatibility.rb
@@ -1,60 +1,6 @@
 require 'pavlov'
 
 module Pavlov
-  def self.old_command command_name, *args
-    klass = class_for_command(command_name)
-    command command_name, arguments_to_attributes(klass, args)
-  end
-
-  def self.old_interactor interactor_name, *args
-    klass = class_for_interactor(interactor_name)
-    interactor interactor_name, arguments_to_attributes(klass, args)
-  end
-
-  def self.old_query query_name, *args
-    klass = class_for_query(query_name)
-    query query_name, arguments_to_attributes(klass, args)
-  end
-
-  def self.arguments_to_attributes(operation_class, arguments)
-    attribute_keys = operation_class.attribute_set.map(&:name)
-
-    # TODO: this can be done so much better, but I don't know how.
-    hash = {}
-    arguments.each_with_index do |value, index|
-      hash[attribute_keys[index].to_sym] = value
-    end
-    hash
-  end
-
-  module Helpers
-    def old_interactor name, *args
-      args = add_compatibility_pavlov_options args
-      Pavlov.old_interactor name, *args
-    end
-
-    def old_query name, *args
-      args = add_compatibility_pavlov_options args
-      Pavlov.old_query name, *args
-    end
-
-    def old_command name, *args
-      args = add_compatibility_pavlov_options args
-      Pavlov.old_command name, *args
-    end
-
-    private
-
-    def add_compatibility_pavlov_options args
-      # TODO: we should do this at a point where we know how many arguments we need
-      # so we can decide if we need to merge with another options object or
-      # just add it.
-      new_args = Array(args)
-      new_args << pavlov_options unless pavlov_options == {}
-      new_args
-    end
-  end
-
   class ValidationError < StandardError
   end