+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
 
+* Removed the methods `old_interactor`, `old_command`, and `old_query` from both Pavlov module and Helpers.
 * In preparation for changes to the validation system, we've changed the name of the method you should implement that performs the validations. This was `valid?`, and it is now `validate`. Like `execute`, this method should not be called directly, instead Pavlov provides a wrapper method called `valid?`. This allows us to add default validations without you having to remember to call `super`.
 
 ## 0.1.6